# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Document APIs",
        "contact": {}
    },
    "paths": {
...
        "/cms/v1/tenants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "List Tenants",
                "parameters": [
                    {
                        "enum": [
                            "enable",
                            "disable"
                        ],
                        "type": "string",
                        "x-enum-varnames": [
                            "DataStatusEnable",
                            "DataStatusDisable"
                        ],
                        "name": "data_status",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "created_at.desc",
                        "name": "sorts",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/db.TenantCmsDto"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Create Tenant",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.TenantCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TenantCmsDto"
                        }
                    }
                }
            }
        },
        "/cms/v1/tenants/{tenant_id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Update Tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenant_id",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/db.TenantCmsData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TenantCmsDto"
                        }
                    }
                }
            }
        },
        "/cms/v1/tenants/{tenant_id}/reset-password": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "Cms"
                ],
                "summary": "Reset Password For Tenant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "tenant_id",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.TenantCmsDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.TenantCmsData": {
            "type": "object",
            "required": [
                "address",
                "data_status",
                "email",
                "keycode",
                "name",
                "phone",
                "username"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Aloha City"
                },
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "email": {
                    "type": "string",
                    "example": "aloha@email.com"
                },
                "keycode": {
                    "type": "string",
                    "example": "aloha"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "phone": {
                    "type": "string",
                    "example": "0973123456"
                },
                "username": {
                    "type": "string",
                    "example": "aloha"
                }
            }
        },
        "db.TenantCmsDto": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Aloha City"
                },
                "data_status": {
                    "$ref": "#/definitions/enum.DataStatus"
                },
                "email": {
                    "type": "string",
                    "example": "aloha@email.com"
                },
                "keycode": {
                    "type": "string",
                    "example": "aloha"
                },
                "name": {
                    "type": "string",
                    "example": "Aloha"
                },
                "phone": {
                    "type": "string",
                    "example": "0973123456"
                },
                "tenant_id": {
                    "type": "string",
                    "example": "671dfc49f06ba89b1811cc5a"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2006-01-02T03:04:05Z"
                },
                "updated_by": {
                    "type": "string",
                    "example": "editor"
                },
                "username": {
                    "type": "string",
                    "example": "aloha"
                }
            }
        },
...
        "enum.DataStatus": {
            "type": "string",
            "enum": [
                "enable",
                "disable"
            ],
            "x-enum-varnames": [
                "DataStatusEnable",
                "DataStatusDisable"
            ]
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "Header"
        }
    }
}